pyproject.toml
[tool.poetry]
name = "bot"
version = "1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.4.1"
python-dotenv = "^0.14.0"
colorama = "^0.4.3"
authlib = "^0.14.3"
httpx = "0.14.1"
aioboto3 = "^8.0.5"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
